requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.1
defusedxml==0.7.1
Django==5.0.9
django-filter==23.5
djangorestframework==3.15.0
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
